src/ai_model_integration.py
```python
import openai

class AIModelIntegration:
    def __init__(self):
        self.ai_model = None

    def integrateAIModel(self, model_name):
        """
        Function to integrate the AI model.
        """
        try:
            self.ai_model = openai.GPT3(model_name)
            print("AI Model integrated successfully.")
        except Exception as e:
            print("Error in integrating AI model: ", str(e))

    def upgradeAIModel(self, new_model_name):
        """
        Function to upgrade the AI model to a more advanced version.
        """
        try:
            self.ai_model = openai.GPT3(new_model_name)
            print("AI Model upgraded successfully.")
        except Exception as e:
            print("Error in upgrading AI model: ", str(e))

    def getAIModel(self):
        """
        Function to get the current AI model.
        """
        return self.ai_model
```