tests/test_ai_model_integration.py
```python
import unittest
from unittest.mock import patch
from src import ai_model_integration

class TestAIModelIntegration(unittest.TestCase):

    @patch('src.ai_model_integration.integrateAIModel')
    def test_integrateAIModel(self, mock_integrateAIModel):
        # Test if the function can integrate a new AI model
        ai_model_integration.integrateAIModel('GPT-4')
        mock_integrateAIModel.assert_called_with('GPT-4')

    @patch('src.ai_model_integration.ai_model')
    def test_ai_model(self, mock_ai_model):
        # Test if the AI model is available and can be accessed
        ai_model_integration.ai_model = 'GPT-3'
        self.assertEqual(ai_model_integration.ai_model, 'GPT-3')

    @patch('src.ai_model_integration.accessAIModel')
    def test_accessAIModel(self, mock_accessAIModel):
        # Test if the function can access and utilize the AI model effectively
        ai_model_integration.accessAIModel('GPT-3')
        mock_accessAIModel.assert_called_with('GPT-3')

if __name__ == '__main__':
    unittest.main()
```